% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myconstr.R
\name{myconstr}
\alias{myconstr}
\title{Constructor Function}
\usage{
myconstr(x, y, alpha)
}
\arguments{
\item{x}{data list}

\item{y}{data list}

\item{alpha}{error level}
}
\value{
a named list containing data, alpha value, confidence interval, p-Value
}
\description{
Constructor Function
}
\examples{
set.seed(21);x <-rnorm(30,5,2); set.seed(23); y<- rnorm(40, 3,2);
alpha <- 0.05; obj <- myconstr(x = x, y = y, alpha = 0.05)

}
