% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myttest.R
\name{myttest}
\alias{myttest}
\title{Constructor Function}
\usage{
myttest(x, y, paired = FALSE, alpha = 0.05)
}
\arguments{
\item{x}{data list}

\item{y}{data list}

\item{paired}{flag for paired t-test}

\item{alpha}{error level}
}
\value{
a named list containing data, alpha value, confidence interval, p-Value
}
\description{
Constructor Function
}
\examples{
set.seed(21);x <-rnorm(30,5,2); set.seed(23); y <- rnorm(40, 3,2);
alpha <- 0.05; obj <- myttest(x = x, y = y, alpha = 0.05)

}
