% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.Rttest}
\alias{plot.Rttest}
\title{Plot Function}
\usage{
\method{plot}{Rttest}(x, ...)
}
\arguments{
\item{x}{data}

\item{...}{others just to make the arguments match the base function}
}
\value{
a box plot
}
\description{
Plot Function
}
